🐛 Fix issue with PRQL build Now, PRQL is not included in dev builds unless the prql tags is specified with make prql
