+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "go.buildTags": "vtable"
+    "go.buildTags": "vtable fts5 sqlite_json sqlite_math_functions"
 }
