+++ other/prql/prql_blank.go
@@ -1,0 +1,7 @@
+//go:build !prql
+
+package prql
+
+func ToSQL(prqlQuery string) (string, []CompileMessage) {
+	return prqlQuery, nil
+}
